Code cleanup: arg to gsx_mfset() is now MFORM *.

+++ aes/gemctrl.c
@@ -315,7 +315,7 @@ void ct_mouse(WORD grabit)
           if (gl_tmpmoff)
             gsx_moff();
           gl_moff = gl_tmpmoff;
-          gsx_mfset((LONG)&gl_mouse);
+          gsx_mfset(&gl_mouse);
           gl_ctmown = FALSE;
           wm_update(FALSE);
         }
